requirements.txt
python-telegram-bot==20.7
google-api-python-client==2.117.0
google-auth==2.29.0
python-dotenv==1.0.1
google-api-python-client
google-auth
google-auth-httplib2
python-dotenv
apscheduler
pytz